Hash tags. Improved command message rendering.

+++ console/controllers/UnsplashDownloadPhotoController.php
@@ -13,12 +13,22 @@ class UnsplashDownloadPhotoController extends Controller
     public function actionIndex()
     {
         $unsplashPhotoList = UnsplashSearchPhoto::find()
-            ->with('setting')
-            ->where(['downloaded_at' => null])
+            ->joinWith(
+                'setting setting',
+                false
+            )
+            ->where([
+                'downloaded_at' => null,
+                'setting.is_active' => true,
+            ])
             ->orderBy(['created_at' => SORT_ASC])
             ->all();
 
+        $this->stdout(sprintf("Count photos: %d\n", count($unsplashPhotoList)), Console::FG_YELLOW);
+
         foreach ($unsplashPhotoList as $key => $unsplashPhoto) {
+            $this->stdout(sprintf("Searched by: %s\n", $unsplashPhoto->setting->search), Console::FG_YELLOW);
+
             $filename = Yii::getAlias("@storage/photoStock/unsplash/{$unsplashPhoto->unsplash_id}.jpg");
 
             $isPhotoDownloaded = file_put_contents(
@@ -26,15 +36,18 @@ class UnsplashDownloadPhotoController extends Controller
                 file_get_contents($unsplashPhoto->raw_url)
             );
 
-            if (false === $isPhotoDownloaded || $isPhotoDownloaded === 0) {
-                $this->stdout("Photo downloading error", Console::FG_RED);
+            if (false === $isPhotoDownloaded || 0 === $isPhotoDownloaded) {
+                $this->stdout("Photo downloading failed", Console::FG_RED);
                 continue;
             }
+
+            $this->stdout("Photo downloaded\n", Console::FG_GREEN);
 
             $socialNetworkPhoto = new SocialNetworkPhoto();
             $socialNetworkPhoto->social_network_account_id = $unsplashPhoto->setting->social_network_account_id;
             $socialNetworkPhoto->filename = $filename;
             $socialNetworkPhoto->file_caption = $unsplashPhoto->description;
+            $socialNetworkPhoto->hash_tags = null;
 
             $unsplashPhoto->downloaded_at = date('Y-m-d H:i:s');
 
@@ -44,16 +57,20 @@ class UnsplashDownloadPhotoController extends Controller
                 $unsplashPhoto->save();
                 $socialNetworkPhoto->save();
 
-                $this->stdout(($key + 1) . " photo downloaded\n", Console::FG_GREEN);
-
                 Yii::$app->db->transaction->commit();
             } catch (\Exception $e) {
                 Yii::$app->db->transaction->rollBack();
 
                 $this->stdout(sprintf("DB error message: %s", $e->getMessage()), Console::FG_RED);
 
-                unlink($filename);
+                if (file_exists($filename)) {
+                    unlink($filename);
+                }
             }
+
+            $this->stdout(" - - - - - \n");
         }
+
+        $this->stdout("Finished!\n", Console::FG_GREEN);
     }
 }
